Add final game() function to finish project
This will provide the user with many
console.log statements to understand what's
going on behind the scenes. It will call all
the other smaller helper functions to make
the game run seamlessly. It will provide an output
of who won the game after all the results have
been calulated. It will also return a message
each round of who won so that the user can
keep up and find out what the computer played.